Fixed reshape/rank problem
lsd_demo.py: Reshaping is tricky and a little bit complex now. I will improve the readability afterward.

frames.py: Normalize should be done uniformly, not columnwise. Foreground masking routine added.

alm_lsd.py, lsd_operations.py: added the functionality that keeps the info of the rank of L.
(I guess we might be able to guess the rank to be always 1 for this application.)